@@ -1,45 +1,44 @@
 {
-  "name": "music",
+  "name": "@pengubot/music",
   "version": "0.0.1",
-  "description": "PenguBot Music Module",
+  "description": "PenguBot Music module",
   "main": "dist/index",
   "types": "dist/index.d.ts",
   "repository": "https://github.com/PenguBot/music.git",
-  "author": "Aditya N. Tripathi <adityatripathidelhi@gmail.com>",
+  "author": "Pengubot",
   "license": "Apache-2.0",
   "private": true,
   "scripts": {
-    "prepublishOnly": "yarn compile",
     "lint": "eslint --fix --ext ts src",
     "test": "eslint --ext ts src",
-    "compile": "tsc -p .",
+    "build": "tsc -p .",
     "watch": "tsc -p . -w",
-    "clean": "rm -rf dist; yarn compile;"
+    "clean": "rm -rf dist; yarn build"
   },
-  "devDependencies": {
-    "@types/cheerio": "^0.22.18",
-    "@types/node": "^14.0.4",
-    "@types/node-fetch": "^2.5.7",
-    "@types/ws": "^7.2.4",
-    "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "@typescript-eslint/parser": "^2.33.0",
-    "discord.js": "github:discordjs/discord.js#fe7df708e44e0280dfaf0f8e457b154781bb5140",
-    "eslint": "^7.0.0",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^4.2.1",
-    "klasa": "github:dirigeants/klasa#be581f178e829b628f38f00d6397a98d27181600",
-    "typescript": "^3.9.3"
-  },
-  "engines": {
-    "node": ">=10"
-  },
-  "files": [
-    "dist"
-  ],
   "dependencies": {
-    "@klasa/utils": "^0.0.5",
+    "@klasa/utils": "^0.0.8",
     "@lavacord/discord.js": "^0.0.7",
     "cheerio": "^1.0.0-rc.3",
     "node-fetch": "^2.6.0"
-  }
+  },
+  "devDependencies": {
+    "@types/cheerio": "^0.22.18",
+    "@types/node": "^14.0.6",
+    "@types/node-fetch": "^2.5.7",
+    "@types/ws": "^7.2.4",
+    "@typescript-eslint/eslint-plugin": "^3.0.2",
+    "@typescript-eslint/parser": "^3.0.2",
+    "klasa": "dirigeants/klasa#settings",
+    "discord.js": "github:discordjs/discord.js#fe7df708e44e0280dfaf0f8e457b154781bb5140",
+    "eslint": "^7.1.0",
+    "eslint-plugin-node": "^11.1.0",
+    "eslint-plugin-promise": "^4.2.1",
+    "typescript": "^3.9.3"
+  },
+  "engines": {
+    "node": ">=12.6"
+  },
+  "files": [
+    "dist"
+  ]
 }
