@@ -1,4 +1,4 @@
-import { Event, EventOptions } from "klasa";
+import { Event, EventOptions, ScheduledTask } from "klasa";
 import { ApplyOptions } from "../lib/utils/Decorators";
 
 @ApplyOptions<EventOptions>({ event: "ready" })
@@ -10,6 +10,13 @@ export default class extends Event {
             for (const node of [...this.client.lavalink.nodes.values()]) if (node.resumeKey) node.resumeKey += `-${shardID}`;
         }
         await this.client.lavalink.connect();
+
+        await this.ensureTask("spotify", "*/30 * * * *");
+    }
+
+    private ensureTask(name: string, time: string): Promise<ScheduledTask> | void {
+        const schedules = this.client.settings!.get("schedules") as any[];
+        if (!schedules.some(task => task.taskName === name)) return this.client.schedule.create(name, time);
     }
 
 }
